stocks/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
dj-static==0.0.6
Django==3.2.19
django-cors-headers==4.0.0
djongo==1.3.6
dnspython==2.3.0
gunicorn==20.1.0
idna==3.4
NorenRestApiPy-0.0.22-py2.py3-none-any.whl
numpy==1.24.3
pandas==2.0.1
pandas-ta==0.3.14b0
pymongo==3.12.1
pyotp==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.30.0
six==1.16.0
sqlparse==0.2.4
static3==0.7.0
tzdata==2023.3
urllib3==2.0.2
websocket-client==1.5.1